backend/requirements.txt
# ============================================
# Schumacher - Backend Dependencies
# Python 3.12+
# ============================================

# ============================================
# FastAPI & Web Framework
# ============================================
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17
websockets<12.0

# ============================================
# Database & ORM
# ============================================
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.31.0  # Async PostgreSQL driver
psycopg2-binary==2.9.10  # Sync PostgreSQL driver (for Alembic)

# ============================================
# Pydantic & Validation
# ============================================
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0

# ============================================
# Redis & Caching
# ============================================
redis==5.2.0
aioredis==2.0.1

# ============================================
# Celery & Background Tasks
# ============================================
celery==5.4.0
flower==2.0.1  # Celery monitoring

# ============================================
# AI & LLM
# ============================================
langchain==0.1.20
langchain-anthropic==0.1.15
langchain-openai==0.1.20
langgraph==0.0.40
anthropic==0.39.0
openai==1.54.4

# LangSmith (optional - for debugging)
langsmith==0.1.143

# ============================================
# Solana SDK
# ============================================
solana==0.34.3
solders==0.21.0
anchorpy==0.20.1

# ============================================
# Ethereum / Web3
# ============================================
web3==7.0.0
eth-keys==0.5.1
eth-typing==4.1.0

# ============================================
# HTTP Clients
# ============================================
httpx==0.27.2
aiohttp==3.11.7
requests==2.32.3

# ============================================
# Authentication & Security
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
cryptography==43.0.3
pynacl==1.5.0  # Ed25519 signature verification (Solana wallets)
python-dotenv==1.0.1

# ============================================
# Monitoring & Logging
# ============================================
sentry-sdk[fastapi]==2.18.0
python-json-logger==3.1.0
structlog==24.4.0

# ============================================
# Utilities
# ============================================
python-dateutil==2.9.0
pytz==2024.2
tenacity==9.0.0  # Retry logic
pydash==8.0.3

# ============================================
# Testing
# ============================================
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0

faker==33.0.0

# ============================================
# Code Quality
# ============================================
ruff==0.7.4  # Linter & formatter (replaces black, isort, flake8)
mypy==1.13.0  # Type checking
pre-commit==4.0.1

# ============================================
# Development Tools
# ============================================
ipython==8.29.0
ipdb==0.13.13
watchdog==6.0.0

# ============================================
# Documentation
# ============================================
mkdocs==1.6.1
mkdocs-material==9.5.47
